change size_t to int for random to work

+++ src/GameCore/map.h
@@ -32,9 +32,9 @@ class Map {
   static constexpr double kVelocityDecrease = 0.75;
   static constexpr double kHPDecrease = 0.001;
   static constexpr size_t kMaxBonusesAmount = 5;
-  static constexpr size_t kAmountOfBonusTypes = 3;
-  static constexpr size_t kMaxMilliSecondsForNewBonus = 20000;
-  static constexpr size_t kMinMilliSecondForNewBonus = 5000;
+  static constexpr int kAmountOfBonusTypes = 3;
+  static constexpr int kMaxMilliSecondsForNewBonus = 20000;
+  static constexpr int kMinMilliSecondForNewBonus = 5000;
 
   void CalculateBonusesPositions();
   void ProceedCollisions(Car*);
